rename article-click-handlers.js to single-article-click-handlers.js
