+++ src/components/SingleArticle/single-article-click-handlers.js
